backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.3",
    "pydantic-ai-slim[openai]>=1.0.17",
    "pydantic-settings>=2.11.0",
    "sqlmodel>=0.0.27",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt>=2.10.1",
    "uuid>=1.30",
    "pytesseract>=0.3.13",
    "pillow>=12.0.0",
    "opencv-python>=4.12.0.88",
    "easyocr>=1.7.2",
    "paddleocr>=3.3.0",
    "doctr>=1.9.0",
    "python-multipart>=0.0.20",
    "pymupdf>=1.26.5",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
    "zuban>=0.1.0",
]
